Starting work on refactoring the tuning API.
Moved global variables in apex_policies.cpp into a new struct which
encapsulates them. Currently there is only one global instance of the struct,
so there's no actual change yet. The next step is to have the functions
in apex_policies.cpp take the tuning_session struct as a parameter.