Split APEX_HAVE_HPX into another macro, APEX_HAVE_HPX_CONFIG
This is annoying and confusing.  We have to set a define so that the
HPX config file will be included, which will define APEX_HAVE_HPX
for us.  We can't use the same name because then the macro is
defined twice.  So, we have a macro to make sure the macro is defined.
This is all because of circular dependencies between HPX and APEX.