Minor gtrace timestamp changes PErfetto seems to have changed how it interprets flow event time stamps. This was detected by testing with the new support for the Iris tasking runtime. For that reason, we now add 0.250 us to the start of the flow event and subtract 0.250 us from the end of the flow event. the timestamps are adjusted to meet their definition of "enclosing slice" and "begin >= timestamp of the flow".
Here is what the documentation says:

```
Each flow event phase has a specific type of binding point:
* Flow start (ph='s'): binding point is always "enclosing slice"
* Flow step (ph='t'): binding point is "enclosing slice" (not used by APEX)
* Flow end (ph='f'): binding point is "next slice". If the event contains
bp="e" then the binding point is "enclosing slice."
A bit of detail on "next slice" binding rules: Next slice is defined as the next slice to begin >= the timestamp of the flow. In the case of multiple slices having the same timestamp as the flow event, the earliest event in the trace buffer is chosen.
```

this commit also includes a change to optionally include timers around
the lifetime of the pthread. it's not really necessary so off by default.